refactor: move DeleteFilesystemModal to the components folder and update import paths
